+++ OpgaveFolder/src/fuzzy/norm/ProbabilisticNorm.java
@@ -1,7 +1,5 @@
 package fuzzy.norm;
 
-import fuzzy.Norm;
-
 /**
  * ProbabilisticNorm - Probabilistic t-(co-)norm pair
  * @author Florian Dejonckheere <florian@floriandejonckheere.be>
@@ -9,12 +7,12 @@ import fuzzy.Norm;
 public class ProbabilisticNorm implements Norm {
 
     @Override
-    public int norm(int a, int b) {
+    public double norm(double a, double b) {
         return (a * b);
     }
 
     @Override
-    public int conorm(int a, int b) {
+    public double conorm(double a, double b) {
         return (a + b - (a * b));
     }
