+++ OpgaveFolder/src/fuzzy/norm/ZadehNorm.java
@@ -1,7 +1,5 @@
 package fuzzy.norm;
 
-import fuzzy.Norm;
-
 /**
  * ZadehNorm - Zadeh t-(co-)norm pair
  * @author Florian Dejonckheere <florian@floriandejonckheere.be>
@@ -9,12 +7,12 @@ import fuzzy.Norm;
 public class ZadehNorm implements Norm {
 
     @Override
-    public int norm(int a, int b) {
+    public double norm(double a, double b) {
         return Math.min(a, b);
     }
 
     @Override
-    public int conorm(int a, int b) {
+    public double conorm(double a, double b) {
         return Math.max(a, b);
     }
